split old lib into GCP core and GCP autocomplete libs
