+++ .eslintrc.js
@@ -14,5 +14,6 @@ module.exports = {
   rules: {
     quotes: ["error", "double"],
     curly: ["error", "all"],
+    "no-use-before-define": ["error", { functions: false, classes: false }],
   },
 };
