+++ gcp-autocomplete-lib.js
@@ -1,31 +1,12 @@
-/* eslint no-use-before-define: 0 */
-
-const _ = require("lodash");
-const { ProjectsClient } = require("@google-cloud/resource-manager");
 const GCCompute = require("@google-cloud/compute");
+const { ProjectsClient } = require("@google-cloud/resource-manager");
+const _ = require("lodash");
+
+const {
+  callMethod,
+} = require("./gcp-lib");
 const parsers = require("./parsers");
 const helpers = require("./helpers");
-
-const RESOURCE_OPERATIONS = {
-  create: _.partial(callMethod, "insert"),
-  get: async (client, request) => (await _.partial(callMethod, "get")(client, request))[0],
-  delete: _.partial(callMethod, "delete"),
-};
-
-async function callResourceOperation(resourceOperation, params, settings, clientClass, resource) {
-  const credentials = helpers.getCredentials(params, settings);
-  const authorizedClient = helpers.getAuthorizedClient(clientClass, credentials);
-
-  const project = helpers.getProject(params, settings);
-  const request = _.merge({ project }, resource);
-
-  const result = await resourceOperation(
-    authorizedClient,
-    request,
-    params.waitForOperation,
-  );
-  return result;
-}
 
 function createListItemsFunction(clientClass, fields) {
   return async function listGcpItems(query, pluginSettings, pluginParams) {
@@ -118,32 +99,7 @@ async function listResources(params, settings, clientClass, resource = {}) {
   return res;
 }
 
-async function callMethod(methodName, client, request, waitForOperation = false) {
-  const result = await client.clientInstance[methodName](request);
-
-  if (!waitForOperation) {
-    return result;
-  }
-
-  const operationsClient = new GCCompute.GlobalOperationsClient({
-    credentials: client.credentials,
-  });
-
-  let [operation] = result;
-  while (operation.status !== "DONE") {
-    // eslint-disable-next-line no-await-in-loop
-    [operation] = await operationsClient.wait({
-      operation: operation.name,
-      project: request.project,
-    });
-  }
-
-  return operation;
-}
-
 module.exports = {
-  RESOURCE_OPERATIONS,
-  callResourceOperation,
   createListItemsFunction,
   listGcpProjects,
   listGcpRegions,
